refactor: modularize serializers and filter views by category, warehouse, price, and attributes
