+++ products/serializers/category_serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from dbmodels.models import Categories
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Categories
+        fields = ['id', 'name']
